requirements.txt
flask
requests
python-dotenv
httpx
sqlalchemy
python-dotenv
rich
